extract rlookup_load_document module
