+++ src/rlookup_load_document.h
@@ -1,0 +1,95 @@
+/*
+ * Copyright (c) 2006-Present, Redis Ltd.
+ * All rights reserved.
+ *
+ * Licensed under your choice of the Redis Source Available License 2.0
+ * (RSALv2); or (b) the Server Side Public License v1 (SSPLv1); or (c) the
+ * GNU Affero General Public License v3 (AGPLv3).
+*/
+#ifndef RLOOKUP_LOAD_DOCUMENT_H
+#define RLOOKUP_LOAD_DOCUMENT_H
+#include <stdint.h>
+#include <assert.h>
+
+#include <spec.h>
+#include <search_ctx.h>
+#include "value.h"
+#include "sortable.h"
+#include "util/arr.h"
+#include "rlookup.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+typedef enum {
+  /* Use keylist (keys/nkeys) for the fields to list */
+  RLOOKUP_LOAD_KEYLIST,
+  /* Load only cached keys (don't open keys) */
+  RLOOKUP_LOAD_SVKEYS,
+  /* Load all keys in the document */
+  RLOOKUP_LOAD_ALLKEYS,
+  /* Load all the keys in the RLookup object */
+  RLOOKUP_LOAD_LKKEYS
+} RLookupLoadFlags;
+
+typedef struct {
+  struct RedisSearchCtx *sctx;
+
+  /** Needed for the key name, and perhaps the sortable */
+  const RSDocumentMetadata *dmd;
+
+  /* Needed for rule filter where dmd does not exist */
+  const char *keyPtr;
+
+  DocumentType type;
+
+  /** Keys to load. If present, then loadNonCached and loadAllFields is ignored */
+  const RLookupKey *const *keys;
+
+  /** Number of keys in keys array */
+  size_t nkeys;
+
+  /**
+   * The following options control the loading of fields, in case non-SORTABLE
+   * fields are desired.
+   */
+  RLookupLoadFlags mode;
+
+  /**
+   * Don't use sortables when loading documents. This will enforce the loader to load
+   * the fields from the document itself, even if they are sortables and un-normalized.
+   */
+  bool forceLoad;
+
+  /**
+   * Force string return; don't coerce to native type
+   */
+  bool forceString;
+
+  struct QueryError *status;
+} RLookupLoadOptions;
+
+/**
+ * Attempt to load a document into the row. The document's fields are placed into
+ * their corresponding slots.
+ *
+ * @param lt Lookup table. Contains the keys to load.
+ * @param dst row that should contain the data
+ * @param options options controlling the load process
+ */
+int RLookup_LoadDocument(RLookup *lt, RLookupRow *dst, RLookupLoadOptions *options);
+
+/**
+ * Initialize the lookup with fields from hash.
+ */
+int RLookup_LoadRuleFields(RedisSearchCtx *sctx, RLookup *it, RLookupRow *dst, IndexSpec *sp, const char *keyptr, QueryError *status);
+
+// exposed to be called from Rust, was inline before that.
+int loadIndividualKeys(RLookup *it, RLookupRow *dst, RLookupLoadOptions *options);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif
