+++ src/spec.c
@@ -47,6 +47,7 @@
 #include "util/redis_mem_info.h"
 #include "search_disk.h"
 #include "search_disk_utils.h"
+#include "rlookup_load_document.h"
 
 #define INITIAL_DOC_TABLE_SIZE 1000
